refactor(server): split HTTP unit tests by handler
Move the 1.4k-line http/tests.rs dump into
tests/{serde_contract,error,codec,sse,handlers,export,admin}.rs.
Shared engine/WAV helpers stay in tests/mod.rs.
